[wallet] rename new wallet module `s/wallet_new/wallet`
