+++ apps/src/lib/client/signing.rs
@@ -5,12 +5,12 @@ use anoma::types::address::Address;
 
 use super::rpc;
 use crate::cli;
-use crate::wallet_new::{self, DecryptedKeypair};
+use crate::wallet::{self, DecryptedKeypair};
 
 /// Find the public key for the given address and try to load the keypair for it
 /// from the wallet. Panics if the key cannot be found or loaded.
 pub async fn find_keypair<'wallet>(
-    wallet: &'wallet wallet_new::Wallet,
+    wallet: &'wallet wallet::Wallet,
     addr: &Address,
     ledger_address: tendermint::net::Address,
 ) -> DecryptedKeypair<'wallet> {
