+++ apps/src/lib/mod.rs
@@ -10,7 +10,7 @@ pub mod logging;
 pub mod node;
 pub mod proto;
 pub mod types;
-pub mod wallet_new;
+pub mod wallet;
 
 // This is here only to include the std's docs in our docs.
 // Taken from <https://github.com/rust-lang/rfcs/issues/2324#issuecomment-502437904>.
