+++ apps/src/lib/wallet/defaults.rs
