+++ apps/src/lib/wallet/store.rs
